remove db config
