+++ pkg/solana/client/client_test.go
@@ -19,7 +19,6 @@ import (
 	"github.com/smartcontractkit/chainlink-common/pkg/logger"
 
 	"github.com/smartcontractkit/chainlink-solana/pkg/solana/config"
-	"github.com/smartcontractkit/chainlink-solana/pkg/solana/db"
 )
 
 func TestClient_Reader_Integration(t *testing.T) {
@@ -31,7 +30,7 @@ func TestClient_Reader_Integration(t *testing.T) {
 
 	requestTimeout := 5 * time.Second
 	lggr := logger.Test(t)
-	cfg := config.NewConfig(db.ChainCfg{}, lggr)
+	cfg := config.NewDefault()
 
 	c, err := NewClient(url, cfg, requestTimeout, lggr)
 	require.NoError(t, err)
@@ -105,7 +104,7 @@ func TestClient_Reader_ChainID(t *testing.T) {
 
 	requestTimeout := 5 * time.Second
 	lggr := logger.Test(t)
-	cfg := config.NewConfig(db.ChainCfg{}, lggr)
+	cfg := config.NewDefault()
 	c, err := NewClient(mockServer.URL, cfg, requestTimeout, lggr)
 	require.NoError(t, err)
 
@@ -126,7 +125,7 @@ func TestClient_Writer_Integration(t *testing.T) {
 
 	requestTimeout := 5 * time.Second
 	lggr := logger.Test(t)
-	cfg := config.NewConfig(db.ChainCfg{}, lggr)
+	cfg := config.NewDefault()
 
 	ctx := context.Background()
 	c, err := NewClient(url, cfg, requestTimeout, lggr)
@@ -212,7 +211,7 @@ func TestClient_SendTxDuplicates_Integration(t *testing.T) {
 	// create client
 	requestTimeout := 5 * time.Second
 	lggr := logger.Test(t)
-	cfg := config.NewConfig(db.ChainCfg{}, lggr)
+	cfg := config.NewDefault()
 	c, err := NewClient(url, cfg, requestTimeout, lggr)
 	require.NoError(t, err)
