+++ pkg/solana/config/config.go
@@ -2,15 +2,11 @@ package config
 
 import (
 	"errors"
-	"strings"
 	"time"
 
 	"github.com/gagliardetto/solana-go/rpc"
 
 	"github.com/smartcontractkit/chainlink-common/pkg/config"
-
-	"github.com/smartcontractkit/chainlink-solana/pkg/solana/db"
-	"github.com/smartcontractkit/chainlink-solana/pkg/solana/logger"
 )
 
 // Global solana defaults.
@@ -73,172 +69,6 @@ type configSet struct {
 	ComputeUnitPriceMin     uint64
 	ComputeUnitPriceDefault uint64
 	FeeBumpPeriod           time.Duration
-}
-
-var _ Config = (*cfg)(nil)
-
-// Deprecated
-type cfg struct {
-	defaults configSet
-	chain    db.ChainCfg
-	lggr     logger.Logger
-}
-
-// NewConfig returns a Config with defaults overridden by dbcfg.
-// Deprecated
-func NewConfig(dbcfg db.ChainCfg, lggr logger.Logger) *cfg {
-	return &cfg{
-		defaults: defaultConfigSet,
-		chain:    dbcfg,
-		lggr:     lggr,
-	}
-}
-
-func (c *cfg) BalancePollPeriod() time.Duration {
-	ch := c.chain.BalancePollPeriod
-	if ch != nil {
-		return ch.Duration()
-	}
-	return c.defaults.BalancePollPeriod
-}
-
-func (c *cfg) ConfirmPollPeriod() time.Duration {
-	ch := c.chain.ConfirmPollPeriod
-	if ch != nil {
-		return ch.Duration()
-	}
-	return c.defaults.ConfirmPollPeriod
-}
-
-func (c *cfg) OCR2CachePollPeriod() time.Duration {
-	ch := c.chain.OCR2CachePollPeriod
-	if ch != nil {
-		return ch.Duration()
-	}
-	return c.defaults.OCR2CachePollPeriod
-}
-
-func (c *cfg) OCR2CacheTTL() time.Duration {
-	ch := c.chain.OCR2CacheTTL
-	if ch != nil {
-		return ch.Duration()
-	}
-	return c.defaults.OCR2CacheTTL
-}
-
-func (c *cfg) TxTimeout() time.Duration {
-	ch := c.chain.TxTimeout
-	if ch != nil {
-		return ch.Duration()
-	}
-	return c.defaults.TxTimeout
-}
-
-func (c *cfg) TxRetryTimeout() time.Duration {
-	ch := c.chain.TxRetryTimeout
-	if ch != nil {
-		return ch.Duration()
-	}
-	return c.defaults.TxRetryTimeout
-}
-
-func (c *cfg) TxConfirmTimeout() time.Duration {
-	ch := c.chain.TxConfirmTimeout
-	if ch != nil {
-		return ch.Duration()
-	}
-	return c.defaults.TxConfirmTimeout
-}
-
-func (c *cfg) SkipPreflight() bool {
-	ch := c.chain.SkipPreflight
-	if ch.Valid {
-		return ch.Bool
-	}
-	return c.defaults.SkipPreflight
-}
-
-func (c *cfg) Commitment() rpc.CommitmentType {
-	ch := c.chain.Commitment
-	if ch.Valid {
-		str := ch.String
-		var commitment rpc.CommitmentType
-		switch str {
-		case "processed":
-			commitment = rpc.CommitmentProcessed
-		case "confirmed":
-			commitment = rpc.CommitmentConfirmed
-		case "finalized":
-			commitment = rpc.CommitmentFinalized
-		default:
-			c.lggr.Warnf(`Invalid value provided for %s, "%s" - falling back to default "%s"`, "CommitmentType", str, c.defaults.Commitment)
-			commitment = rpc.CommitmentConfirmed
-		}
-		return commitment
-	}
-	return c.defaults.Commitment
-}
-
-func (c *cfg) FeeEstimatorMode() string {
-	ch := c.chain.FeeEstimatorMode
-	if ch.Valid {
-		return strings.ToLower(ch.String)
-	}
-	return c.defaults.FeeEstimatorMode
-}
-
-func (c *cfg) ComputeUnitPriceMax() uint64 {
-	ch := c.chain.ComputeUnitPriceMax
-	if ch.Valid {
-		if ch.Int64 >= 0 {
-			return uint64(ch.Int64)
-		}
-		c.lggr.Warnf("Negative value provided for ComputeUnitPriceMax, falling back to default: %d", c.defaults.ComputeUnitPriceMax)
-	}
-	return c.defaults.ComputeUnitPriceMax
-}
-
-func (c *cfg) ComputeUnitPriceMin() uint64 {
-	ch := c.chain.ComputeUnitPriceMin
-	if ch.Valid {
-		if ch.Int64 >= 0 {
-			return uint64(ch.Int64)
-		}
-		c.lggr.Warnf("Negative value provided for ComputeUnitPriceMin, falling back to default: %d", c.defaults.ComputeUnitPriceMin)
-	}
-	return c.defaults.ComputeUnitPriceMin
-}
-
-func (c *cfg) ComputeUnitPriceDefault() uint64 {
-	ch := c.chain.ComputeUnitPriceDefault
-	if ch.Valid {
-		if ch.Int64 >= 0 {
-			return uint64(ch.Int64)
-		}
-		c.lggr.Warnf("Negative value provided for ComputeUnitPriceDefault, falling back to default: %d", c.defaults.ComputeUnitPriceDefault)
-	}
-	return c.defaults.ComputeUnitPriceDefault
-}
-
-func (c *cfg) MaxRetries() *uint {
-	ch := c.chain.MaxRetries
-	if ch.Valid {
-		if ch.Int64 < 0 {
-			c.lggr.Warnf(`Negative value provided for %s: %d, falling back to <nil> - let RPC node do a reasonable amount of tries`, "MaxRetries", ch.Int64)
-			return nil
-		}
-		val := uint(ch.Int64)
-		return &val
-	}
-	return c.defaults.MaxRetries
-}
-
-func (c *cfg) FeeBumpPeriod() time.Duration {
-	ch := c.chain.FeeBumpPeriod
-	if ch != nil {
-		return ch.Duration()
-	}
-	return c.defaults.FeeBumpPeriod
 }
 
 type Chain struct {
